Refactors messaging library packaging
